Refactor: service 패키지에서 repository 패지키로 이동
